backend/requirements.txt

Flask==2.2.2
Flask-Cors==3.0.10
trimesh==3.23.5
numpy==1.23.5
scipy==1.10.0
Pillow==9.4.0
